src/test/java/state/PlayerTest.java
package state;

import org.junit.jupiter.api.Test;
import stonegames.model.Player;

import java.util.Date;

import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.junit.jupiter.api.Assertions.assertTrue;

public class PlayerTest {

    @Test
    public void testPlayer() throws InterruptedException {
        Player player=new Player();
        Date date1=new Date();
        player.setStartTime(date1);
        player.setPlayerName("playertest");
        player.setCount(100);
        Thread.sleep(5000);
        Date date2=new Date();
        player.setEndTime(date2);
        player.setScore();


        assertNotNull(player);
        assertTrue("playertest".equals(player.getPlayerName()));
        assertTrue(100==player.getCount());
        assertTrue(date1.equals(player.getStartTime()));
        assertTrue(date2.equals(player.getEndTime()));
        assertTrue(1000000/((int)(((date2.getTime()-date1.getTime())/1000)*100+100))==player.getScore());
    }


}
